+++ projects/zenvdb/MeshToSDF.cpp
@@ -130,18 +130,18 @@ struct SDFToPoly : zeno::INode{
     mesh->resize(points.size());
     mesh->add_attr<zeno::vec3f>("pos");
 #pragma omp parallel for
-    for(size_t i=0;i<points.size();i++)
+    for(int i=0;i<points.size();i++)
     {
         mesh->attr<zeno::vec3f>("pos")[i] = zeno::vec3f(points[i][0],points[i][1],points[i][2]);
     }
     mesh->tris.resize(tris.size() + 2*quads.size());
 #pragma omp parallel for
-    for(size_t i=0;i<tris.size();i++)
+    for(int i=0;i<tris.size();i++)
     {
         mesh->tris[i] = zeno::vec3i(tris[i][0],tris[i][1],tris[i][2]);
     }
 #pragma omp parallel for
-    for(size_t i=0;i<quads.size();i++)
+    for(int i=0;i<quads.size();i++)
     {
         mesh->tris[i*2+tris.size()] = zeno::vec3i(quads[i][0],quads[i][1],quads[i][2]);
         mesh->tris[i*2+1+tris.size()] = zeno::vec3i(quads[i][2],quads[i][3],quads[i][0]);
